Move OS into it's own file
